Big commit. Main change: rewritten VMM
After many attempts making the most basic VMM, the first baby step
is finally complete (not even close honestly ^^). kalloc() is able
to allocated the amount of requested bytes and returns a void
pointer which is VA. It's use is in kernel's area only (3-4GB).
But the algorithm is still totally broken and allocated memory
cannot be freed.
This commit also changed all tabs to spaces and other small
fixes.